Weather Service: add simulated data and time clock multiplier (#33)
* feat(weatherService): add the weather service data with csv and clock time multiplier

* feat(dataManager): add the dataManager class that handles the data collection from the dataset and to give them to the services

* feat(energyPrice): call the dataManageer for the energy price to get the data from the dataset

+++ src/Interfaces/IClock.hpp
@@ -4,6 +4,7 @@
  */
 
 #pragma once
+#include <cstdint>
 
 namespace POLA::Interfaces {
 
@@ -12,8 +13,8 @@ class IClock
 public:
     virtual ~IClock() = default;
     virtual void simulate() = 0;
-    [[nodiscard]] virtual double getElapsedTimeSinceStart() const = 0;
-    [[nodiscard]] virtual double getElapsedTime() const = 0;
+    [[nodiscard]] virtual uint32_t getElapsedTimeSinceStart() const = 0;
+    [[nodiscard]] virtual uint32_t getElapsedTime() const = 0;
 };
 
 } // namespace POLA::Interfaces
